refactor: extracting project config
eliminate unnecessary default values, explicit error ignoring, project interface for future facade testing, putting schemas within the dedicated packages (no schema package!), moving project config from internal, typos
