+++ pkg/project/config_test.go
@@ -1,4 +1,4 @@
-package schema
+package project
 
 import (
 	"io/ioutil"
@@ -14,12 +14,14 @@ func TestCreateDefaultProjectConfig(t *testing.T) {
 		t.Fatalf("can't create temporary directory: %s", err)
 	}
 
-	defaultProjectConfig := NewDefaultProjectConfig()
+	defaultProjectConfig := newDefaultProjectConfig()
 
 	// create temporary file to test error on existing file
 	existingKlioFileName := "existing-klio.yaml"
 	existingKlioFileNameAbsPath := path.Join(dir, existingKlioFileName)
-	defer os.RemoveAll(existingKlioFileNameAbsPath)
+	defer func(path string) {
+		_ = os.RemoveAll(path)
+	}(existingKlioFileNameAbsPath)
 
 	_, err = os.Create(path.Join(dir, existingKlioFileName))
 	if err != nil {
@@ -32,7 +34,7 @@ func TestCreateDefaultProjectConfig(t *testing.T) {
 	tests := []struct {
 		name    string
 		args    args
-		want    *ProjectConfig
+		want    *config
 		wantErr bool
 	}{
 		{
@@ -55,11 +57,13 @@ func TestCreateDefaultProjectConfig(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			filePath := path.Join(dir, tt.args.filePath)
-			defer os.Remove(filePath)
+			defer func(name string) {
+				_ = os.Remove(name)
+			}(filePath)
 
-			_, err := CreateDefaultProjectConfig(filePath)
+			_, err := CreateDefaultConfig(filePath)
 			if (err != nil) != tt.wantErr {
-				t.Errorf("CreateDefaultProjectConfig() error = %v, wantErr %v", err, tt.wantErr)
+				t.Errorf("CreateDefaultConfig() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
